+++ src/test/java/mappify/special/TestObjectWeirdKeys.java
@@ -10,7 +10,6 @@ public class TestObjectWeirdKeys {
     @Save
     @Typing({WeirdKey.class, WeirdKey.class})
     public Map<WeirdKey,WeirdKey> map = new HashMap<>();
-
     {
         map.put(new WeirdKey("weird"),new WeirdKey("sweared"));
         map.put(new WeirdKey("beard"),new WeirdKey("smeared"));