+++ CHANGELOG.md
@@ -41,3 +41,7 @@ Initializers will be called when unmappifying objects instead of calling Field s
 # 1.4.5 - Now uploading directly to a public maven separate from github packages
 
 # 1.4.6 - Fix a fatal bug with Transform keys due to missing array index.
+
+# 1.4.7 - Added @Priority and @Ordering both of which allow further control over field saving order.
+Also fixed a bug where classes containing required static and non-static fields would not be able to be deserialized properly.
+WrappedObjectMap now extends LinkedHashMap so class level ordering should be maintained now.